# MINOR: Unnecessary changeCapacity in ImplicitLinkedHashCollection