1. remove ImplicitLinkedHashCollection.add unnecessary changeCapacity
2. remove ImplicitLinkedHashCollection.changeCapacity unnecessary iterator.remove

#### trunk

<img width="1312" alt="image" src="https://github.com/user-attachments/assets/f44be3b1-22dd-4a18-bcb5-fdb627711a60">

```
Benchmark                                                                (size)  Mode  Cnt       Score       Error  Units
ImplicitLinkedHashCollectionBenchmark.testCreateFromEmpty                 10000  avgt    6    4397.636 ±   326.099  us/op
ImplicitLinkedHashCollectionBenchmark.testCreateFromEmpty                100000  avgt    6  767246.324 ± 45883.873  us/op
ImplicitLinkedHashCollectionBenchmark.testCreateFromExpectedNumElements   10000  avgt    6     814.625 ±   421.052  us/op
ImplicitLinkedHashCollectionBenchmark.testCreateFromExpectedNumElements  100000  avgt    6   34003.380 ±  2006.759  us/op
```

#### this pr
```
Benchmark                                                                (size)  Mode  Cnt      Score      Error  Units
ImplicitLinkedHashCollectionBenchmark.testCreateFromEmpty                 10000  avgt    6    523.661 ±  251.641  us/op
ImplicitLinkedHashCollectionBenchmark.testCreateFromEmpty                100000  avgt    6  22689.922 ± 1312.356  us/op
ImplicitLinkedHashCollectionBenchmark.testCreateFromExpectedNumElements   10000  avgt    6    262.020 ±   24.802  us/op
ImplicitLinkedHashCollectionBenchmark.testCreateFromExpectedNumElements  100000  avgt    6   2306.306 ±  253.645  us/op
```